+++ bank/src/main/java/com/baas/bank/BankApplication.java
@@ -6,6 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 @MapperScan("com.baas.bank.account.mapper")
+		@MapperScan("com.baas.bank.user.mapper"), // 사용자 매퍼 경로
 public class BankApplication {
 
 	public static void main(String[] args) {
