+++ bank/src/main/java/com/baas/bank/user/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.baas.bank.account.dto;
+package com.baas.bank.user.dto;
 
 import lombok.*;
