+++ bank/src/main/java/com/baas/bank/user/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.baas.bank.account.exception;
+package com.baas.bank.user.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
