+++ bank/src/main/java/com/baas/bank/user/exception/InvalidCredentialsException.java
@@ -1,4 +1,4 @@
-package com.baas.bank.account.exception;
+package com.baas.bank.user.exception;
 
 public class InvalidCredentialsException extends RuntimeException {
     public InvalidCredentialsException(String message) {
