+++ bank/src/main/java/com/baas/bank/user/exception/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.baas.bank.account.exception;
+package com.baas.bank.user.exception;
 
 public class UserAlreadyExistsException extends RuntimeException {
     public UserAlreadyExistsException(String message) {
