+++ bank/src/main/resources/mapper/UserMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.baas.bank.account.mapper.UserMapper">
+<mapper namespace="com.baas.bank.user.mapper.UserMapper">
 
     <insert id="insertUser" parameterType="UserDto">
         INSERT INTO user (login_id, login_password, name, birth_date, phone, email)
